+++ python/cudf/tests/test_dataframe.py
@@ -83,6 +83,10 @@ def test_series_basic():
     assert len(series) == 10
     np.testing.assert_equal(series.to_array(), np.hstack([a1]))
 
+
+def test_series_append():
+    a1 = np.arange(10, dtype=np.float64)
+    series = Series(a1)
     # Add new buffer
     a2 = np.arange(5)
     series = series.append(a2)
@@ -95,6 +99,14 @@ def test_series_basic():
     assert len(series) == 18
     a4 = np.hstack([a1, a2, a3])
     np.testing.assert_equal(series.to_array(), a4)
+
+    # Appending different dtype
+    a5 = np.array([1, 2, 3], dtype=np.int32)
+    a6 = np.array([4.5, 5.5, 6.5], dtype=np.float64)
+    series = Series(a5).append(a6)
+    np.testing.assert_equal(series.to_array(), np.hstack([a5, a6]))
+    series = Series(a6).append(a5)
+    np.testing.assert_equal(series.to_array(), np.hstack([a6, a5]))
 
 
 def test_series_indexing():
@@ -1110,3 +1122,14 @@ def test_dataframe_boolean_mask_Series(gdf):
     assert gdf_masked.shape[0] == 2
     assert gdf_masked2.shape[0] == 4
     assert gdf_masked3.shape[0] == 8
+
+
+def test_iter(pdf, gdf):
+    assert list(pdf) == list(gdf)
+
+
+def test_iteritems(gdf):
+    for k, v in gdf.iteritems():
+        assert k in gdf.columns
+        assert isinstance(v, gd.Series)
+        assert_eq(v, gdf[k])